requirements.txt
python-telegram-bot==20.8
requests
urllib3
nest_asyncio
